+++ packages/blip/app/components/chart/daily.js
@@ -26,6 +26,7 @@ import { TimeService } from 'medical-domain'
 import { components as vizComponents } from 'tidepool-viz'
 import Footer from './footer'
 import {
+  AlarmEventTooltip,
   BloodGlucoseTooltip,
   BolusTooltip,
   ConfidentialTooltip,
@@ -72,6 +73,7 @@ class DailyChart extends React.Component {
     onPhysicalHover: PropTypes.func.isRequired,
     onParameterHover: PropTypes.func.isRequired,
     onWarmUpHover: PropTypes.func.isRequired,
+    onAlarmEventHover: PropTypes.func.isRequired,
     onConfidentialHover: PropTypes.func.isRequired,
     onTooltipOut: PropTypes.func.isRequired,
     onChartMounted: PropTypes.func.isRequired,
@@ -94,6 +96,7 @@ class DailyChart extends React.Component {
       'onParameterHover',
       'onConfidentialHover',
       'onWarmUpHover',
+      'onAlarmEventHover',
       'onTooltipOut',
       'trackMetric'
     ]
@@ -335,7 +338,7 @@ class Daily extends React.Component {
 
     return (
       <div id="tidelineMain" className="daily">
-        <Box className="container-box-outer patient-data-content-outer" display="flex" flexDirection="column">
+        <Box data-testid="daily-view-content" className="container-box-outer patient-data-content-outer" display="flex" flexDirection="column">
           <Box display="flex">
             {this.state.chartMounted &&
               <DailyDatePicker
@@ -380,6 +383,7 @@ class Daily extends React.Component {
                   onPhysicalHover={this.handlePhysicalHover}
                   onParameterHover={this.handleParameterHover}
                   onWarmUpHover={this.handleWarmUpHover}
+                  onAlarmEventHover={this.handleAlarmEventHover}
                   onConfidentialHover={this.handleConfidentialHover}
                   onTooltipOut={this.handleTooltipOut}
                   onChartMounted={this.onChartMounted}
@@ -640,6 +644,22 @@ class Daily extends React.Component {
     this.setState({ tooltip })
   }
 
+  handleAlarmEventHover = (datum) => {
+    this.updateDatumHoverForTooltip(datum)
+    const tooltip = (
+      <AlarmEventTooltip
+        alarmEvent={datum.data}
+        position={{
+          top: datum.top,
+          left: datum.left
+        }}
+        side={datum.side}
+        bgPrefs={datum.bgPrefs}
+        timePrefs={datum.timePrefs}
+      />)
+    this.setState({ tooltip })
+  }
+
   handleConfidentialHover = (datum) => {
     this.updateDatumHoverForTooltip(datum)
     const tooltip = (
@@ -657,7 +677,7 @@ class Daily extends React.Component {
 
   /**
    * Update the daily view by adding the new message
-   * @param {object} message A nurseshark processed message
+   * @param {object} message A processed message
    * @return {Promise<boolean>} true if the message was added
    */
   createMessage = (message) => {
@@ -666,7 +686,7 @@ class Daily extends React.Component {
 
   /**
    * Update the daily view message
-   * @param {object} message A nurseshark processed message
+   * @param {object} message A processed message
    * @return {boolean} true if the message was correctly updated
    */
   editMessage = (message) => {