+++ packages/medical-domain/src/domains/models/statistics/carbs-statistics.model.ts
@@ -25,8 +25,21 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import { HoursRange } from './satistics.model'
+
 export interface CarbsStatistics {
   totalCarbsPerDay: number
-  foodCarbsPerDay: number
-  totalEntriesCarbWithRescueCarbs: number
+  rescueCarbsPerDay: number
+  mealCarbsPerDay: number
+  estimatedCarbsPerDay: number
+  totalMealCarbsWithRescueCarbsEntries: number
+  totalRescueCarbsEntries: number
 }
+
+export interface RescueCarbsAveragePerRange {
+  numberOfIntakes: number
+  confirmedCarbs: number
+  recommendedCarbs: number
+}
+
+export type RescueCarbsAverageStatistics = Map<HoursRange, RescueCarbsAveragePerRange>