+++ packages/medical-domain/src/domains/repositories/medical/medical-data.service.ts
@@ -73,9 +73,13 @@ import type PumpSettings from '../../models/medical/datum/pump-settings.model'
 import { DatumType } from '../../models/medical/datum/enums/datum-type.enum'
 import type PumpManufacturer from '../../models/medical/datum/enums/pump-manufacturer.enum'
 import WizardService from './datum/wizard.service'
+import { DeviceEventSubtype } from '../../models/medical/datum/enums/device-event-subtype.enum';
+import { AlarmEvent } from '../../models/medical/datum/alarm-event.model';
+import { AlarmEventType } from '../../models/medical/datum/enums/alarm-event-type.enum';
 
 class MedicalDataService {
   medicalData: MedicalData = {
+    alarmEvents: [],
     basal: [],
     bolus: [],
     cbg: [],
@@ -110,17 +114,12 @@ class MedicalDataService {
   }
 
   public get latestPumpManufacturer(): string {
-    const capitalize = (s: string): string => {
-      const lower = s.toLowerCase()
-      return lower[0].toUpperCase() + lower.slice(1)
-    }
     const pumpSettings = this.medicalData.pumpSettings
     if (pumpSettings.length > 0) {
       const latestPumpSettings = this.getLatestPumpSettings(pumpSettings)
-      const manufacturer = latestPumpSettings.payload.pump.manufacturer.toLowerCase()
-      return capitalize(manufacturer)
+      return latestPumpSettings.payload.pump.manufacturer
     } else {
-      return capitalize(this._datumOpts.defaultPumpManufacturer)
+      return this._datumOpts.defaultPumpManufacturer
     }
   }
 
@@ -289,7 +288,7 @@ class MedicalDataService {
     rawData.forEach(raw => {
       try {
         const datum = DatumService.normalize(raw, this._datumOpts)
-        const deviceEventDatum = datum as ConfidentialMode | DeviceParameterChange | ReservoirChange | WarmUp | ZenMode
+        const deviceEventDatum = datum as AlarmEvent | ConfidentialMode | DeviceParameterChange | ReservoirChange | WarmUp | ZenMode
 
         switch (datum.type) {
           case DatumType.Bolus:
@@ -303,19 +302,25 @@ class MedicalDataService {
             break
           case DatumType.DeviceEvent:
             switch (deviceEventDatum.subType as string) {
-              case 'confidential':
+              case DeviceEventSubtype.Alarm:
+                const alarmEvent = deviceEventDatum as AlarmEvent
+                if (alarmEvent.alarmEventType !== AlarmEventType.Unknown) {
+                  this.medicalData.alarmEvents.push(alarmEvent)
+                }
+                break
+              case DeviceEventSubtype.Confidential:
                 this.medicalData.confidentialModes.push(deviceEventDatum as ConfidentialMode)
                 break
-              case 'deviceParameter':
+              case DeviceEventSubtype.DeviceParameter:
                 this.medicalData.deviceParametersChanges.push(deviceEventDatum as DeviceParameterChange)
                 break
-              case 'reservoirChange':
+              case DeviceEventSubtype.ReservoirChange:
                 this.medicalData.reservoirChanges.push(deviceEventDatum as ReservoirChange)
                 break
-              case 'warmup':
+              case DeviceEventSubtype.Warmup:
                 this.medicalData.warmUps.push(deviceEventDatum as WarmUp)
                 break
-              case 'zen':
+              case DeviceEventSubtype.Zen:
                 this.medicalData.zenModes.push(deviceEventDatum as ZenMode)
                 break
               default: