+++ packages/yourloops/test/integration/setup.js
@@ -25,12 +25,15 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import '@testing-library/jest-dom/extend-expect'
+import '@testing-library/jest-dom'
 import { enableFetchMocks } from 'jest-fetch-mock'
 
-/* This is done because the chat widget is using the scroll method, not implemented by jsdom which is manipulating only dom no layout */
+// Mocking the `scroll` method as it is not implemented in jsdom (used by the Chat widget component)
 Element.prototype.scroll = jest.fn()
 window.scroll = jest.fn()
+// Mocking the `scrollIntoView` method as it is not implemented in jsdom (used by the Target & alerts page)
+Element.prototype.scrollIntoView = jest.fn()
+
 global.URL.createObjectURL = jest.fn().mockReturnValue('/')
 
 enableFetchMocks()